revert: undo util.py and index_support.py splits
Restores util.py (802 lines) and index_support.py (623 lines) to their
original single-file contents and removes util_ranges.py,
util_textfiles.py, and index_formats.py. Test references and the
critique/report notes for those splits are reverted too. The
geometry_support package split and the hermetic test suite (93.1%
coverage) are unaffected.
